old/python/requirements.txt
requests==2.18.4
websocket-client==0.46.0
